+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## _(in-progress)_
+## 1.8.3 _(2021-05-01)_
 * Fix: Change `__wakeup()` method visibility from `private` to `public` to avoid warnings under PHP8
 * Fix: Throw an error when attempting to unserialize an instance of the class to actually prevent it from happening